Use "VideoView.setUser" to set the user name and avatar from "webrtc.js"
If the user is a guest and no name is given "Guest" is set by default;
this causes a minor behaviour difference in behaviour when the user data
is set after creating the view: before, the name was not shown until the
connection was made, but now "Guest" is shown for guests (and only for
guests) as soon as the view is created. This will be fixed once the
names are got from the participant list instead of the WebRTC
connection.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>